feat: merge rss3 workers to vsl core
