+++ internal/engine/worker/decentralized/core/ethereum/worker.go
@@ -18,6 +18,7 @@ import (
 	"github.com/rss3-network/node/provider/ethereum/contract/erc20"
 	"github.com/rss3-network/node/provider/ethereum/contract/erc721"
 	"github.com/rss3-network/node/provider/ethereum/contract/rss3"
+	"github.com/rss3-network/node/provider/ethereum/contract/vsl"
 	"github.com/rss3-network/node/provider/ethereum/token"
 	workerx "github.com/rss3-network/node/schema/worker/decentralized"
 	"github.com/rss3-network/protocol-go/schema"
@@ -36,14 +37,16 @@ import (
 var _ engine.Worker = (*worker)(nil)
 
 type worker struct {
-	config             *config.Module
-	ethereumClient     ethereum.Client
-	tokenClient        token.Client
-	erc20Filterer      *erc20.ERC20Filterer
-	erc721Filterer     *erc721.ERC721Filterer
-	erc1155Filterer    *erc1155.ERC1155Filterer
-	stakingVSLFilterer *rss3.StakingVSLFilterer
-	chipsFilterer      *rss3.ChipsFilterer
+	config                           *config.Module
+	ethereumClient                   ethereum.Client
+	tokenClient                      token.Client
+	erc20Filterer                    *erc20.ERC20Filterer
+	erc721Filterer                   *erc721.ERC721Filterer
+	erc1155Filterer                  *erc1155.ERC1155Filterer
+	stakingVSLFilterer               *rss3.StakingVSLFilterer
+	chipsFilterer                    *rss3.ChipsFilterer
+	contractL1StandardBridgeFilterer *vsl.L1StandardBridgeFilterer
+	contractL2StandardBridgeFilterer *vsl.L2StandardBridgeFilterer
 }
 
 func (w *worker) Name() string {
@@ -86,6 +89,7 @@ func (w *worker) Types() []schema.Type {
 		typex.TransactionApproval,
 		typex.TransactionMint,
 		typex.TransactionBurn,
+		typex.TransactionBridge,
 		typex.CollectibleTransfer,
 		typex.CollectibleApproval,
 		typex.CollectibleMint,
@@ -156,7 +160,26 @@ func (w *worker) Transform(ctx context.Context, task engine.Task) (*activityx.Ac
 			)
 
 			switch {
-			//	VSL Staking
+			// VSL Bridge
+			case w.matchL1StandardBridgeETHDepositInitiatedLog(ethereumTask, log):
+				logActions, err = w.transformL1StandardBridgeETHDepositInitiatedLog(ctx, ethereumTask, log)
+				activity.Platform = workerx.PlatformVSL.String()
+			case w.matchL1StandardBridgeERC20DepositInitiatedLog(ethereumTask, log):
+				logActions, err = w.transformL1StandardBridgeERC20DepositInitiatedLog(ctx, ethereumTask, log)
+				activity.Platform = workerx.PlatformVSL.String()
+			case w.matchL1StandardBridgeETHWithdrawalFinalizedLog(ethereumTask, log):
+				logActions, err = w.transformL1StandardBridgeETHWithdrawalFinalizedLog(ctx, ethereumTask, log)
+				activity.Platform = workerx.PlatformVSL.String()
+			case w.matchL1StandardBridgeERC20WithdrawalFinalizedLog(ethereumTask, log):
+				logActions, err = w.transformL1StandardBridgeERC20WithdrawalFinalizedLog(ctx, ethereumTask, log)
+				activity.Platform = workerx.PlatformVSL.String()
+			case w.matchL2StandardBridgeWithdrawalInitiatedLog(ethereumTask, log):
+				logActions, err = w.transformL2StandardBridgeWithdrawalInitiatedLog(ctx, ethereumTask, log)
+				activity.Platform = workerx.PlatformVSL.String()
+			case w.matchL2StandardBridgeDepositFinalizedLog(ethereumTask, log):
+				logActions, err = w.transformL2StandardBridgeDepositFinalizedLog(ctx, ethereumTask, log)
+				activity.Platform = workerx.PlatformVSL.String()
+			// VSL Staking
 			case w.matchStakingVSLDeposited(ethereumTask, log):
 				logActions, err = w.handleStakingVSLDeposited(ctx, ethereumTask, log)
 				activity.Platform = workerx.PlatformVSL.String()
@@ -253,6 +276,30 @@ func (w *worker) matchStakingVSLStaked(_ *source.Task, log *ethereum.Log) bool {
 // matchTransfer matches the transfer event.
 func (w *worker) matchChipsTransfer(_ *source.Task, log *ethereum.Log) bool {
 	return log.Address == rss3.AddressChipsVSL && contract.MatchEventHashes(log.Topics[0], rss3.EventHashTransfer)
+}
+
+func (w *worker) matchL1StandardBridgeETHDepositInitiatedLog(_ *source.Task, log *ethereum.Log) bool {
+	return log.Address == vsl.AddressL1StandardBridge && log.Topics[0] == vsl.EventHashAddressL1StandardBridgeETHDepositInitiated
+}
+
+func (w *worker) matchL1StandardBridgeERC20DepositInitiatedLog(_ *source.Task, log *ethereum.Log) bool {
+	return log.Address == vsl.AddressL1StandardBridge && log.Topics[0] == vsl.EventHashAddressL1StandardBridgeERC20DepositInitiated
+}
+
+func (w *worker) matchL1StandardBridgeETHWithdrawalFinalizedLog(_ *source.Task, log *ethereum.Log) bool {
+	return log.Address == vsl.AddressL1StandardBridge && log.Topics[0] == vsl.EventHashAddressL1StandardBridgeETHWithdrawalFinalized
+}
+
+func (w *worker) matchL1StandardBridgeERC20WithdrawalFinalizedLog(_ *source.Task, log *ethereum.Log) bool {
+	return log.Address == vsl.AddressL1StandardBridge && log.Topics[0] == vsl.EventHashAddressL1StandardBridgeERC20WithdrawalFinalized
+}
+
+func (w *worker) matchL2StandardBridgeWithdrawalInitiatedLog(_ *source.Task, log *ethereum.Log) bool {
+	return log.Address == vsl.AddressL2StandardBridge && log.Topics[0] == vsl.EventHashAddressL2StandardBridgeWithdrawalInitiated
+}
+
+func (w *worker) matchL2StandardBridgeDepositFinalizedLog(_ *source.Task, log *ethereum.Log) bool {
+	return log.Address == vsl.AddressL2StandardBridge && log.Topics[0] == vsl.EventHashAddressL2StandardBridgeDepositFinalized
 }
 
 func (w *worker) handleNativeTransferTransaction(ctx context.Context, task *source.Task) (*activityx.Action, error) {
@@ -478,6 +525,114 @@ func (w *worker) handleChipsMint(ctx context.Context, task *source.Task, log *et
 	}, nil
 }
 
+func (w *worker) transformL1StandardBridgeETHDepositInitiatedLog(ctx context.Context, task *source.Task, log *ethereum.Log) ([]*activityx.Action, error) {
+	event, err := w.contractL1StandardBridgeFilterer.ParseETHDepositInitiated(log.Export())
+	if err != nil {
+		return nil, fmt.Errorf("parse ETHDepositInitiated event: %w", err)
+	}
+
+	action, err := w.buildTransactionBridgeAction(ctx, task.ChainID, event.From, event.To, network.Ethereum, network.VSL, metadata.ActionTransactionBridgeDeposit, nil, event.Amount, log.BlockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("build transaction bridge action: %w", err)
+	}
+
+	actions := []*activityx.Action{
+		action,
+	}
+
+	return actions, nil
+}
+
+func (w *worker) transformL1StandardBridgeERC20DepositInitiatedLog(ctx context.Context, task *source.Task, log *ethereum.Log) ([]*activityx.Action, error) {
+	event, err := w.contractL1StandardBridgeFilterer.ParseERC20DepositInitiated(log.Export())
+	if err != nil {
+		return nil, fmt.Errorf("parse ERC20DepositInitiated event: %w", err)
+	}
+
+	action, err := w.buildTransactionBridgeAction(ctx, task.ChainID, event.From, event.To, network.Ethereum, network.VSL, metadata.ActionTransactionBridgeDeposit, &event.L1Token, event.Amount, log.BlockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("build transaction bridge action: %w", err)
+	}
+
+	actions := []*activityx.Action{
+		action,
+	}
+
+	return actions, nil
+}
+
+func (w *worker) transformL1StandardBridgeETHWithdrawalFinalizedLog(ctx context.Context, task *source.Task, log *ethereum.Log) ([]*activityx.Action, error) {
+	event, err := w.contractL1StandardBridgeFilterer.ParseETHWithdrawalFinalized(log.Export())
+	if err != nil {
+		return nil, fmt.Errorf("parse ETHWithdrawalFinalized event: %w", err)
+	}
+
+	action, err := w.buildTransactionBridgeAction(ctx, task.ChainID, event.From, event.To, network.VSL, network.Ethereum, metadata.ActionTransactionBridgeWithdraw, nil, event.Amount, log.BlockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("build transaction bridge action: %w", err)
+	}
+
+	actions := []*activityx.Action{
+		action,
+	}
+
+	return actions, nil
+}
+
+func (w *worker) transformL1StandardBridgeERC20WithdrawalFinalizedLog(ctx context.Context, task *source.Task, log *ethereum.Log) ([]*activityx.Action, error) {
+	event, err := w.contractL1StandardBridgeFilterer.ParseERC20WithdrawalFinalized(log.Export())
+	if err != nil {
+		return nil, fmt.Errorf("parse ERC20WithdrawalFinalized event: %w", err)
+	}
+
+	action, err := w.buildTransactionBridgeAction(ctx, task.ChainID, event.From, event.To, network.VSL, network.Ethereum, metadata.ActionTransactionBridgeWithdraw, &event.L1Token, event.Amount, log.BlockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("build transaction bridge action: %w", err)
+	}
+
+	actions := []*activityx.Action{
+		action,
+	}
+
+	return actions, nil
+}
+
+func (w *worker) transformL2StandardBridgeWithdrawalInitiatedLog(ctx context.Context, task *source.Task, log *ethereum.Log) ([]*activityx.Action, error) {
+	event, err := w.contractL2StandardBridgeFilterer.ParseWithdrawalInitiated(log.Export())
+	if err != nil {
+		return nil, fmt.Errorf("parse WithdrawalInitiated event: %w", err)
+	}
+
+	action, err := w.buildTransactionBridgeAction(ctx, task.ChainID, event.From, event.To, network.VSL, network.Ethereum, metadata.ActionTransactionBridgeDeposit, &event.L2Token, event.Amount, log.BlockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("build transaction bridge action: %w", err)
+	}
+
+	actions := []*activityx.Action{
+		action,
+	}
+
+	return actions, nil
+}
+
+func (w *worker) transformL2StandardBridgeDepositFinalizedLog(ctx context.Context, task *source.Task, log *ethereum.Log) ([]*activityx.Action, error) {
+	event, err := w.contractL2StandardBridgeFilterer.ParseDepositFinalized(log.Export())
+	if err != nil {
+		return nil, fmt.Errorf("parse DepositFinalized event: %w", err)
+	}
+
+	action, err := w.buildTransactionBridgeAction(ctx, task.ChainID, event.From, event.To, network.Ethereum, network.VSL, metadata.ActionTransactionBridgeWithdraw, &event.L2Token, event.Amount, log.BlockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("build transaction bridge action: %w", err)
+	}
+
+	actions := []*activityx.Action{
+		action,
+	}
+
+	return actions, nil
+}
+
 func (w *worker) buildTransactionTransferAction(ctx context.Context, task *source.Task, from, to common.Address, tokenAddress *common.Address, tokenValue *big.Int) (*activityx.Action, error) {
 	chainID, err := network.EthereumChainIDString(task.GetNetwork().String())
 	if err != nil {
@@ -657,6 +812,30 @@ func (w *worker) buildChipsMintAction(ctx context.Context, task *source.Task, fr
 	}, nil
 }
 
+func (w *worker) buildTransactionBridgeAction(ctx context.Context, chainID uint64, sender, receiver common.Address, source, target network.Network, bridgeAction metadata.TransactionBridgeAction, tokenAddress *common.Address, tokenValue *big.Int, blockNumber *big.Int) (*activityx.Action, error) {
+	tokenMetadata, err := w.tokenClient.Lookup(ctx, chainID, tokenAddress, nil, blockNumber)
+	if err != nil {
+		return nil, fmt.Errorf("lookup token %s: %w", tokenAddress, err)
+	}
+
+	tokenMetadata.Value = lo.ToPtr(decimal.NewFromBigInt(tokenValue, 0))
+
+	action := activityx.Action{
+		Type:     typex.TransactionBridge,
+		Platform: workerx.PlatformVSL.String(),
+		From:     sender.String(),
+		To:       receiver.String(),
+		Metadata: metadata.TransactionBridge{
+			Action:        bridgeAction,
+			SourceNetwork: source,
+			TargetNetwork: target,
+			Token:         *tokenMetadata,
+		},
+	}
+
+	return &action, nil
+}
+
 // Handle invalid token error.
 func isInvalidTokenErr(err error) bool {
 	return err != nil && strings.Contains(err.Error(), "unsupported NFT standard")
@@ -679,6 +858,13 @@ func NewWorker(config *config.Module, redisClient rueidis.Client) (engine.Worker
 	instance.erc721Filterer = lo.Must(erc721.NewERC721Filterer(ethereum.AddressGenesis, nil))
 	instance.erc1155Filterer = lo.Must(erc1155.NewERC1155Filterer(ethereum.AddressGenesis, nil))
 
+	// Initialize VSL contract filterers.
+	if instance.config.Network == network.Ethereum {
+		instance.contractL1StandardBridgeFilterer = lo.Must(vsl.NewL1StandardBridgeFilterer(ethereum.AddressGenesis, nil))
+		instance.contractL2StandardBridgeFilterer = lo.Must(vsl.NewL2StandardBridgeFilterer(ethereum.AddressGenesis, nil))
+	}
+
+	// Initialize RSS3 contract filterers.
 	if instance.config.Network == network.VSL {
 		instance.stakingVSLFilterer = lo.Must(rss3.NewStakingVSLFilterer(ethereum.AddressGenesis, nil))
 		instance.chipsFilterer = lo.Must(rss3.NewChipsFilterer(ethereum.AddressGenesis, nil))
